requirements.txt
# Core Django
Django>=4.2.7,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-environ>=0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
django-allauth==0.57.0
django-otp==1.2.2
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
django-ratelimit==4.1.0

# SMS Services
twilio==8.10.0
messagebird==2.2.0

# Static files & Media
whitenoise==6.6.0
Pillow==10.1.0

# API & Serialization
djangorestframework-simplejwt==5.3.0
django-filter==23.4
drf-spectacular==0.26.5

# Development & Testing
django-debug-toolbar>=4.2.0
factory-boy>=3.3.0
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0

# Production Server
gunicorn>=21.2.0
celery>=5.3.4
redis>=5.0.1

# Code Quality
flake8==6.1.0
black==24.4.2
isort==5.12.0
bandit==1.7.5
safety>=3.0.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-extensions==3.2.3

# Bootstrap & Frontend
django-bootstrap5==23.3
crispy-bootstrap5==0.7

# Time & Localization
pytz==2023.3
django-timezone-field==6.0.1
